requirements.txt
absl-py
mock
ntplib
pyfakefs
PyYAML
requests
git+git://github.com/google/winops@master#egg=gwinpy
